# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AutoHaul Connect API - direct car transport marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user profile by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List all open shipment listings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, assigned, in_transit, delivered, cancelled]
        - name: originState
          in: query
          schema:
            type: string
        - name: destinationState
          in: query
          schema:
            type: string
        - name: vehicleType
          in: query
          schema:
            type: string
        - name: transportType
          in: query
          schema:
            type: string
            enum: [open, enclosed]
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentListResponse"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a new shipment listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentBody"
      responses:
        "201":
          description: Created shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "401":
          description: Unauthorized
  /shipments/{shipmentId}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment by ID
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "404":
          description: Not found
    put:
      operationId: updateShipment
      tags: [shipments]
      summary: Update a shipment listing
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentBody"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
    delete:
      operationId: deleteShipment
      tags: [shipments]
      summary: Cancel/delete a shipment
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/user/{userId}:
    get:
      operationId: getUserReviews
      tags: [reviews]
      summary: Get reviews for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        authId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [shipper, driver, both, admin]
        profileImageUrl:
          type: string
        bio:
          type: string
        dotNumber:
          type: string
        mcNumber:
          type: string
        insuranceProvider:
          type: string
        insurancePolicyNumber:
          type: string
        truckType:
          type: string
        truckCapacity:
          type: number
        isVerified:
          type: boolean
        isSuspended:
          type: boolean
        averageRating:
          type: number
        totalReviews:
          type: number
        completedJobs:
          type: number
        termsAccepted:
          type: boolean
        termsAcceptedAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - role
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [shipper, driver, both]
        bio:
          type: string
        dotNumber:
          type: string
        mcNumber:
          type: string
        insuranceProvider:
          type: string
        insurancePolicyNumber:
          type: string
        truckType:
          type: string
        truckCapacity:
          type: number
        termsAccepted:
          type: boolean
    Shipment:
      type: object
      properties:
        id:
          type: string
        shipperId:
          type: string
        shipper:
          $ref: "#/components/schemas/UserProfile"
        title:
          type: string
        vehicleYear:
          type: number
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleType:
          type: string
          enum: [sedan, suv, truck, van, motorcycle, rv, exotic, other]
        vehicleCondition:
          type: string
          enum: [running, non_running]
        vin:
          type: string
        transportType:
          type: string
          enum: [open, enclosed]
        originAddress:
          type: string
        originCity:
          type: string
        originState:
          type: string
        originZip:
          type: string
        destinationAddress:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        destinationZip:
          type: string
        pickupDateFrom:
          type: string
        pickupDateTo:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [open, assigned, in_transit, delivered, cancelled]
        bidCount:
          type: number
        assignedDriverId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - shipperId
        - vehicleYear
        - vehicleMake
        - vehicleModel
        - vehicleType
        - vehicleCondition
        - transportType
        - originCity
        - originState
        - destinationCity
        - destinationState
        - status
    CreateShipmentBody:
      type: object
      properties:
        vehicleYear:
          type: number
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleType:
          type: string
          enum: [sedan, suv, truck, van, motorcycle, rv, exotic, other]
        vehicleCondition:
          type: string
          enum: [running, non_running]
        vin:
          type: string
        transportType:
          type: string
          enum: [open, enclosed]
        originAddress:
          type: string
        originCity:
          type: string
        originState:
          type: string
        originZip:
          type: string
        destinationAddress:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        destinationZip:
          type: string
        pickupDateFrom:
          type: string
        pickupDateTo:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        notes:
          type: string
      required:
        - vehicleYear
        - vehicleMake
        - vehicleModel
        - vehicleType
        - vehicleCondition
        - transportType
        - originCity
        - originState
        - originZip
        - destinationCity
        - destinationState
        - destinationZip
    UpdateShipmentBody:
      type: object
      properties:
        notes:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        pickupDateFrom:
          type: string
        pickupDateTo:
          type: string
        status:
          type: string
          enum: [open, cancelled]
    ShipmentListResponse:
      type: object
      properties:
        shipments:
          type: array
          items:
            $ref: "#/components/schemas/Shipment"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - shipments
        - total
...
    Review:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        reviewerId:
          type: string
        reviewerName:
          type: string
        revieweeId:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - reviewerId
        - revieweeId
        - rating
    CreateReviewBody:
      type: object
      properties:
        bookingId:
          type: string
        revieweeId:
          type: string
        rating:
          type: number
        comment:
          type: string
      required:
        - bookingId
        - revieweeId
        - rating
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        averageRating:
          type: number
        total:
          type: number
      required:
        - reviews
        - total
...